scenario-2-ticket-triage/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
asyncpg==0.29.0
pydantic==2.8.2
prometheus-client==0.20.0
python-json-logger==2.0.7
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
google-generativeai==0.7.2
tenacity==8.5.0
PyYAML==6.0.2
requests==2.32.3
